+++ issuer-application/src/main/java/ch/admin/bj/swiyu/issuer/infrastructure/web/management/CredentialManagementController.java
@@ -20,7 +20,6 @@ import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Map;
 import java.util.UUID;
 
 @RestController